% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_all_apps}
\alias{deploy_all_apps}
\title{Wrapper to deploy all teal sample apps at once}
\usage{
deploy_all_apps(
  api_key,
  api_secret = "",
  server,
  title_prefix = "",
  title_suffix = format(Sys.Date(), format = "\%Y_\%m_\%d"),
  vanity_url_prefix = "/NEST",
  create_manifest = FALSE
)
}
\arguments{
\item{api_key}{(\code{character(1)}) API key/token.}

\item{api_secret}{(\code{character(1)}) If deploying to a shinyapps.io account, the token's secret.}

\item{server}{(\code{character(1)}) The URL for accessing the Posit Connect or shinyapps.io server.}

\item{title_prefix}{(\code{character(1)}) Prefix to add to all app titles.}

\item{title_suffix}{(\code{character(1)}) Suffix to add to all app titles.}

\item{vanity_url_prefix}{(\code{character(1)}) Prefix to add to the apps' vanity URLs.}

\item{create_manifest}{(\code{logical(1)}) Whether to create the manifest.json for the app.}
}
\description{
Deploys all teal sample apps at once.
}
\examples{
\dontrun{
teal.gallery::deploy_all_apps()
}
}
