% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_app}
\alias{deploy_app}
\title{Deploy an app from the gallery to any Posit Connect server}
\usage{
deploy_app(
  app_name,
  app_title,
  api_key,
  vanity_url,
  server,
  overwrite = TRUE,
  api_secret = "",
  create_manifest = TRUE
)
}
\arguments{
\item{app_name}{(\code{character(1)}) name of the the teal app - technically, name of the app
directory in the \verb{inst/}. Do not confuse with the \code{name} argument in the \code{\link[connectapi:deploy]{connectapi::deploy()}} .}

\item{app_title}{(\code{character(1)}, optional) The title to be used for the content on the server.
Unique app name deployed on the \code{server} is generated from \code{app_title} by removing}

\item{api_key}{(\code{character(1)}) API key/token.}

\item{vanity_url}{(\code{character(1)}) The path component of the URL where app will be available.}

\item{server}{(\code{character(1)}) The URL for accessing the Posit Connect server.}

\item{overwrite}{(\code{logical(1)}) whether remote app should be overwritten by the current deployment.}

\item{api_secret}{(\code{character(1)}) If deploying to a shinyapps.io account, the token's secret.}

\item{create_manifest}{(\code{logical(1)}) Whether to create the manifest.json for the app.}
}
\description{
Not a lot of flexibility in configuring the app
as the options adhere to a certain set of requirements
}
\examples{
\dontrun{
teal.gallery::deploy_app(
  app_name = "basic-teal",
  app_title = "Basic Teal App - TEST",
  vanity_url = "/NEST/main/basic-teal",
  api_key = "SHINYAPPS_OR_POSIT_CONNECT_API_TOKEN",
  overwrite = FALSE
)
}
}
