% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_app_code}
\alias{get_app_code}
\title{Get the source code of an app}
\usage{
get_app_code(app_name)
}
\arguments{
\item{app_name}{\code{character}: the name of the app. Run \code{list_apps()} to see the
names of the apps in the gallery}
}
\value{
(\code{character}) The code for the chosen app
}
\description{
Get the source code of an app
}
\examples{
if (interactive()) {
  get_app_code("basic-teal")

  # code can be saved into a file
  code <- get_app_code("basic-teal")
  # writeLines(code, con = file("app.R"))
}
}
