% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{launch_app}
\alias{launch_app}
\title{Run an app from the gallery}
\usage{
launch_app(app_name, ...)
}
\arguments{
\item{app_name}{\code{character}: the name of the app. Run \code{list_apps()} to see the
names of the apps in the gallery}

\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}}
}
\description{
Run an app from the gallery
}
\details{
The package uses \code{\link[renv:dependencies]{renv::dependencies()}} to check for
dependencies the app requires and if they are not installed then
the app will not load and an error will list the required packages to run the app
}
\examples{
\dontrun{
launch_app("basic-teal")
}
}
