% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_apps}
\alias{list_apps}
\title{List the \code{teal} apps in the gallery}
\usage{
list_apps()
}
\value{
vector of app names in the gallery
}
\description{
List the \code{teal} apps in the gallery
}
\examples{
if (interactive()) {
  list_apps()
}
}
