% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.gallery.R
\docType{package}
\name{teal.gallery-package}
\alias{teal.gallery}
\alias{teal.gallery-package}
\title{\code{teal.gallery} Package}
\description{
\code{teal.gallery} is a gallery of sample apps based on the Teal framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/teal.gallery}
  \item Report bugs at \url{https://github.com/insightsengineering/teal.gallery/issues}
}

}
\author{
\strong{Maintainer}: Dony Unardi \email{unardid@gene.com}

Authors:
\itemize{
  \item Nikolas Burkoff \email{nikolas.burkoff@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
